move send-qr code in members page
